extracting and splitting history_builder
